refactor(backend): extract common enum
